ref!: Move files to src folder for easier package imports
