+++ lib/src/enum.dart
