+++ lib/src/http_client_adapter.dart
