+++ lib/src/models/exception.dart
