+++ lib/src/models/reconnect.dart
