+++ test/browser/fetch_client_extension_test.dart
@@ -1,11 +1,6 @@
 @TestOn('browser')
-import 'package:eventflux/extensions/fetch_client_extension.dart';
-import 'package:eventflux/models/web_config/redirect_policy.dart';
-import 'package:eventflux/models/web_config/request_cache.dart';
-import 'package:eventflux/models/web_config/request_credentials.dart';
-import 'package:eventflux/models/web_config/request_mode.dart';
-import 'package:eventflux/models/web_config/request_referrer_policy.dart';
-import 'package:eventflux/models/web_config/web_config.dart';
+import 'package:eventflux/src/extensions/fetch_client_extension.dart';
+import 'package:eventflux/src/models/web_config/web_config.dart';
 import 'package:fetch_client/fetch_client.dart';
 import 'package:test/test.dart';
