+++ test/mocks.dart
@@ -1,4 +1,4 @@
-import 'package:eventflux/http_client_adapter.dart';
+import 'package:eventflux/src/http_client_adapter.dart';
 import 'package:mockito/annotations.dart';
 
 @GenerateMocks([HttpClientAdapter])
