+++ test/mocks.mocks.dart
@@ -5,7 +5,7 @@
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'dart:async' as _i4;
 
-import 'package:eventflux/http_client_adapter.dart' as _i3;
+import 'package:eventflux/src/http_client_adapter.dart' as _i3;
 import 'package:http/http.dart' as _i2;
 import 'package:mockito/mockito.dart' as _i1;
