+++ test/models/web_config/web_config_test.dart
@@ -1,9 +1,4 @@
-import 'package:eventflux/models/web_config/redirect_policy.dart';
-import 'package:eventflux/models/web_config/request_cache.dart';
-import 'package:eventflux/models/web_config/request_credentials.dart';
-import 'package:eventflux/models/web_config/request_mode.dart';
-import 'package:eventflux/models/web_config/request_referrer_policy.dart';
-import 'package:eventflux/models/web_config/web_config.dart';
+import 'package:eventflux/src/models/web_config/web_config.dart';
 import 'package:test/test.dart';
 
 void main() {
